+++ collective/auditlog/tasks.py
@@ -4,14 +4,14 @@ from collective.celery import task
 logger = get_task_logger(__name__)
 
 
-@task(name="audit")
+@task(name="collective.auditlog.audit")
 def queue_job(obj, *args, **kwargs):
     from collective.auditlog.async_svc import runJob
     logger.warn('Logging action on {}'.format(obj.absolute_url_path()))
     runJob(obj, *args, **kwargs)
 
 
-@task(name="catalog-entry")
+@task(name="collective.auditlog.catalog-entry")
 def queue_catalog_entry(obj, data):
     from collective.auditlog.catalog import catalogEntry
     logger.warn('Cataloging event for {}'.format(obj.absolute_url_path()))